requirements-gpu.txt
absl-py==0.2.2
astor==0.6.2
bleach==1.5.0
cycler==0.10.0
dill==0.2.7.1
gast==0.2.0
grpcio==1.12.0
h5py==2.7.1
html5lib==0.9999999
Keras==2.1.6
kiwisolver==1.0.1
Markdown==2.6.11
matplotlib==2.2.2
multiprocess==0.70.5
numpy==1.14.3
opencv-python==3.4.1.15
pandas==0.23.0
pathos==0.2.1
Pillow==5.1.0
pox==0.2.3
ppft==1.6.4.7.1
protobuf==3.5.2.post1
pyparsing==2.2.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
scikit-learn==0.19.1
scipy==1.1.0
six==1.11.0
sklearn==0.0
tensorboard==1.8.0
tensorflow-gpu==1.8.0
termcolor==1.1.0
Werkzeug==0.14.1
